+++ source/Gui/StyleRepository.h
@@ -12,6 +12,8 @@ namespace Const
 
     ImColor const RenderingDisabledTextColor = ImColor::HSV(0.5f, 0.1f, 1.0f, 0.2f);
 
+    ImColor const TooltipTextColor = ImColor::HSV(0.0f, 0.0f, 1.0f, 1.0f);
+
     int64_t const SimulationSliderColor_Base = 0xff4c4c4c;
     int64_t const SimulationSliderColor_Active = 0xff6c6c6c;
     int64_t const TextDecentColor = 0xff909090;